moved the header of the page in app.js to a component and added a link to the header to go back to the home page
