+++ client/src/App.js
@@ -9,17 +9,12 @@ import Signup from './pages/Signup';
 import SingleProduct from './pages/SingleProduct';
 import Footer from "./components/Footer";
 import Nav from "./components/Nav";
+import Header from './components/Header'
 
-import { Grid, Container, AppBar, Toolbar, Typography } from "@mui/material";
-import { FormControl, MenuItem, TextField, InputLabel } from "@mui/material";
+import {Container} from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
 import { CssBaseline } from "@mui/material";
-import { useDispatch, useSelector} from 'react-redux';
-import { CHANGE_PAGE_THEME } from "./utils/actions";
-import CoffeeTheme from "./themes/coffee";
-import MintTheme from "./themes/mint";
-import IceTheme from "./themes/ice";
-import DarkBrewTheme from "./themes/dark";
+import {useSelector} from 'react-redux';
 
 
 const httpLink = createHttpLink({
@@ -42,90 +37,30 @@ const httpLink = createHttpLink({
   });
 
 function App() {
-    const dispatch = useDispatch()
     const state = useSelector(state => state)
-
-
-    const handleChange = (event) => {
-        const newTheme = event.target.value
-
-        switch(newTheme){
-            case MintTheme:
-                return dispatch({
-                    type: CHANGE_PAGE_THEME,
-                    theme: MintTheme
-                });
-            case CoffeeTheme:
-                return dispatch({
-                    type: CHANGE_PAGE_THEME,
-                    theme: CoffeeTheme
-                });
-            case IceTheme:
-                return dispatch({
-                    type: CHANGE_PAGE_THEME,
-                    theme: IceTheme
-                });
-            case DarkBrewTheme:
-                return dispatch({
-                    type: CHANGE_PAGE_THEME,
-                    theme: DarkBrewTheme
-                });
-            default:
-                return dispatch({
-                type: CHANGE_PAGE_THEME,
-                theme: CoffeeTheme
-            })
-        }
-
-    };
+    console.log(state)
 
     return (
-        <ThemeProvider theme={state.theme}>
-            <CssBaseline />
-            <Container maxWidth="xl">
-                <Grid container spacing={2} sx={{ mb:2 }}>
-                    <AppBar>
-                        <Toolbar>
-                            <Typography variant="h2" component="div" sx={{ flexGrow: 1 }}>
-                                The Coffee Shop
-                            </Typography>
-                            <FormControl variant="outlined" sx={{ width: "15%", mr: 6 }}>
-                            <InputLabel htmlFor="select-theme" shrink sx={{ color: "success.main" }}>Choose Theme:</InputLabel>
-                                <TextField
-                                    id="select-theme"
-                                    value={state.theme}
-                                    onChange={handleChange}
-                                    select
-                                    variant="filled"
-                                    margin="none"
-                                    size="small"
-                                    SelectProps={{ sx: { color: "success.main"} }}
-                                >
-                                    <MenuItem value={CoffeeTheme} id="coffee">Creamy Coffee</MenuItem>
-                                    <MenuItem value={MintTheme} id="mint">Minty Mocha</MenuItem>
-                                    <MenuItem value={IceTheme} id="ice">Icy Coffee</MenuItem>
-                                    <MenuItem value={DarkBrewTheme} id="ice">Dark Brew</MenuItem>
-                                </TextField>
-                            </FormControl>
-                        </Toolbar>
-                    </AppBar>
-                </Grid>
-                <ApolloProvider client={client}>
-                    <Router>
-                        <div>
-                            <Nav />
-                            <Switch>
-                                <Route exact path="/" component={Home} />
-                                <Route exact path="/login" component={Login} />
-                                <Route exact path="/signup" component={Signup} />
-                                <Route exact path="/products/:id" component={SingleProduct} />
-                            </Switch>
-                        </div>
-                    </Router>
-                </ApolloProvider>
-                <Footer />
-            </Container>
-        </ThemeProvider>
+    <ApolloProvider client={client}>
+        <Router>
+            <ThemeProvider theme={state.theme}>
+                <CssBaseline />
+                <Container maxWidth="xl">
+                    <Header/>
+                    <div>
+                        <Nav />
+                        <Switch>
+                            <Route exact path="/" component={Home} />
+                            <Route  path="/login" component={Login} />
+                            <Route  path="/signup" component={Signup} />
+                            <Route  path="/products/:id" component={SingleProduct} />
+                        </Switch>
+                    </div>
+                    <Footer />
+                </Container>
+            </ThemeProvider>
+        </Router>
+    </ApolloProvider>
     );
 }
